Add files via upload
An Entity class is created which is made an abstract and the Creature class and the Player class will extend the Entity class. Entity is class that will store all the elements of the game which aren't in the background or are in the tiles. These will store objects and temporary characters(NPC) etc.